Move AdaGrad update policy to ada_grad
